+++ src/KmerAnalysis.cpp
@@ -61,17 +61,17 @@ vector<KmerItem_t> CreateKmerVecFromReadSeq(int len, char* seq)
 
 	tail = 0; count = 0;
 
-	while (count < KmerSize && tail < len)
+	while (count < (uint32_t)KmerSize && tail < (uint32_t)len)
 	{
 		if (seq[tail++] != 'N') count++;
 		else count = 0;
 	}
-	if (count == KmerSize) // found the first kmer
+	if (count == (uint32_t)KmerSize) // found the first kmer
 	{
 		WordPosition.pos = (head = tail - KmerSize); WordPosition.wid = CreateKmerID(seq, head);
 		vec.push_back(WordPosition);
 
-		for (head += 1; tail < len; head++, tail++)
+		for (head += 1; tail < (uint32_t)len; head++, tail++)
 		{
 			if (seq[tail] != 'N')
 			{
@@ -83,12 +83,12 @@ vector<KmerItem_t> CreateKmerVecFromReadSeq(int len, char* seq)
 			{
 				// find next kmer without 'N'
 				count = 0; tail++;
-				while (count < KmerSize && tail < len)
+				while (count < (uint32_t)KmerSize && tail < (uint32_t)len)
 				{
 					if (seq[tail++] != 'N') count++;
 					else count = 0;
 				}
-				if (count == KmerSize)
+				if (count == (uint32_t)KmerSize)
 				{
 					WordPosition.pos = (head = tail - KmerSize); WordPosition.wid = CreateKmerID(seq, head);
 					vec.push_back(WordPosition);
@@ -115,7 +115,7 @@ vector<KmerPair_t> IdentifyCommonKmers(int MaxShift, vector<KmerItem_t>& vec1, v
 		wid = vec1[i].wid; KmerPtr = lower_bound(vec2.begin(), vec2.end(), vec1[i], CompByKmerID);
 		while (KmerPtr != vec2.end() && KmerPtr->wid == wid)
 		{
-			if ((KmerPtr->pos >= vec1[i].pos && KmerPtr->pos - vec1[i].pos < MaxShift) || ((KmerPtr->pos < vec1[i].pos && vec1[i].pos - KmerPtr->pos < MaxShift)))
+			if ((KmerPtr->pos >= vec1[i].pos && KmerPtr->pos - vec1[i].pos < (uint32_t)MaxShift) || ((KmerPtr->pos < vec1[i].pos && vec1[i].pos - KmerPtr->pos < (uint32_t)MaxShift)))
 			{
 				KmerPair.rPos = vec1[i].pos;
 				KmerPair.PosDiff = (KmerPair.gPos = KmerPtr->pos) - KmerPair.rPos;
@@ -143,7 +143,7 @@ vector<SeedPair_t> GenerateSimplePairsFromCommonKmers(int MinSeedLength, vector<
 		//printf("KmerPair %d: %d PosDiff = %d\n", i+1, KmerPairVec[i].rPos, KmerPairVec[i].PosDiff);
 		for (PosDiff = KmerPairVec[i].PosDiff, n_pos = KmerPairVec[i].rPos + 1, j = i + 1; j < num; j++)
 		{
-			if (KmerPairVec[j].rPos != n_pos || KmerPairVec[j].PosDiff != PosDiff) break;
+			if (KmerPairVec[j].rPos != (uint32_t)n_pos || KmerPairVec[j].PosDiff != PosDiff) break;
 			//printf("KmerPair: %d PosDiff = %d\n", KmerPairVec[j].rPos, KmerPairVec[j].PosDiff);
 			n_pos++;
 		}
